Extracted HealthStoreProtocol
